report truncation on waf only
